+++ webui/src/lib/components/auth/forms.jsx
@@ -9,15 +9,8 @@ import {SearchIcon} from "@primer/octicons-react";
 import {useAPI} from "../../hooks/api";
 import {Checkbox, DataTable, DebouncedFormControl, AlertError, Loading} from "../controls";
 
-const resolveEntityDisplayName = (ent) => {
-    // for users
-    if (ent?.email?.length) return ent.email;
-    // for groups
-    if (ent?.name?.length) return ent.name;
-    return ent.id;
-}
 
-export const AttachModal = ({ show, searchFn, onAttach, onHide, addText = "Add",
+export const AttachModal = ({ show, searchFn, resolveEntityFn = (ent => ent.id), onAttach, onHide , addText = "Add",
                           emptyState = 'No matches', modalTitle = 'Add', headers = ['Select', 'ID'],
                      filterPlaceholder = 'Filter...'}) => {
     const search = useRef(null);
@@ -49,7 +42,7 @@ export const AttachModal = ({ show, searchFn, onAttach, onHide, addText = "Add",
                             onAdd={() => setSelected([...selected, ent])}
                             onRemove={() => setSelected(selected.filter(selectedEnt => selectedEnt.id !== ent.id))}
                             name={'selected'}/>,
-                        <strong>{resolveEntityDisplayName(ent)}</strong>
+                        <strong>{resolveEntityFn(ent)}</strong>
                     ]}/>
 
                 <div className="mt-3">
@@ -58,7 +51,7 @@ export const AttachModal = ({ show, searchFn, onAttach, onHide, addText = "Add",
                         <strong>Selected: </strong>
                         {(selected.map(item => (
                             <Badge key={item.id} pill variant="primary" className="me-1">
-                                {resolveEntityDisplayName(item)}
+                                {resolveEntityFn(item)}
                             </Badge>
                         )))}
                     </p>